The feature to add favorites has been improved but there are some bugs.

+++ app/src/main/java/com/knbrgns/isparkappclone/repository/ParkRepository.kt
@@ -17,9 +17,10 @@ class ParkRepository {
     suspend fun getParks(): Result<List<Park>> {
         return try {
             val response = parkAPI.getParks()
-
             if (response.isSuccessful) {
-                Result.success(response.body() ?: emptyList())
+                val parks = response.body() ?: emptyList()
+                val updatedParks = checkFavorites(parks)
+                Result.success(updatedParks)
             } else {
                 Result.failure(Exception("Park verileri alınamadı"))
             }
@@ -27,6 +28,15 @@ class ParkRepository {
             Result.failure(e)
         }
     }
+    private suspend fun checkFavorites(parks: List<Park>): List<Park> {
+        val favoriteParks = favoriteDb.getFavoriteParks()
+        val favoriteIds = favoriteParks.map { it.parkID }.toSet()
+
+        return parks.map { park ->
+            park.copy(isFavorite = favoriteIds.contains(park.parkID))
+        }
+    }
+
 
     suspend fun getParkWithId(parkId: Int): Result<Park> {
         return try {
@@ -43,25 +53,20 @@ class ParkRepository {
         }
     }
 
-    suspend fun addToFavorites(park: Park): Result<Unit> {
+    suspend fun toggleFavorite(park: Park): Result<Boolean> {
         return try {
-            favoriteDb.addFavoritePark(park)
-            Result.success(Unit)
+            if (park.isFavorite) {
+                favoriteDb.deleteFavoritePark(park)
+            } else {
+                favoriteDb.addFavoritePark(park)
+            }
+            Result.success(!park.isFavorite)
         } catch (e: Exception) {
             Result.failure(e)
         }
     }
 
-    suspend fun removeFromFavorites(park: Park): Result<Unit> {
-        return try {
-            favoriteDb.deleteFavoritePark(park)
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    suspend fun getFavoriteIds(): Result<List<Park>> {
+    suspend fun getFavoriteParks(): Result<List<Park>> {
         return try {
             val favoriteParks = favoriteDb.getFavoriteParks()
             Result.success(favoriteParks)
